# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Moderation queue counts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/reports:
    get:
      operationId: listAdminReports
      tags: [admin]
      summary: List moderation reports with filters
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: reportType
          in: query
          required: false
          schema:
            type: string
        - name: targetType
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReportList"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/reports/{id}:
    get:
      operationId: getAdminReport
      tags: [admin]
      summary: Full moderation report with reported content
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReportDetail"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/reports/{id}/resolve:
    post:
      operationId: resolveAdminReport
      tags: [admin]
      summary: Update a report's triage state / record an action
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveReportRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/flags:
    get:
      operationId: listAdminFlags
      tags: [admin]
      summary: List auto-flagged accounts
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFlagList"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/flags/{id}/review:
    post:
      operationId: reviewAdminFlag
      tags: [admin]
      summary: Mark an account flag reviewed or dismissed
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewFlagRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{userId}/suspend:
    post:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend a user (temporary or indefinite)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendUserRequest"
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{userId}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Permanently ban a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserRequest"
      responses:
        "200":
          description: Banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{userId}/lift:
    post:
      operationId: liftUserModeration
      tags: [admin]
      summary: Clear a user's suspension/ban
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/photos/{photoId}:
    delete:
      operationId: adminRemovePhoto
      tags: [admin]
      summary: Remove a reported photo
      security:
        - bearerAuth: []
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PublicProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        bio:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        city:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        distance_km:
          type: number
          nullable: true
        is_online:
          type: boolean
        is_verified:
          type: boolean
        liked_by_me:
          type: boolean
        blocked_by_me:
          type: boolean
        plan:
          type: string
          enum: [free, plus, gold]
          nullable: true
          description: Entitlement tier, used for the Gold priority badge.
        created_at:
          type: string
      required: [id, username]
...
    ProfilePhoto:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        url:
          type: string
        storage_path:
          type: string
        is_avatar:
          type: boolean
        position:
          type: integer
        created_at:
          type: string
      required: [id, user_id, url, storage_path, is_avatar, position]
...
    Message:
      type: object
      properties:
        id:
          type: string
        conversation_id:
          type: string
        sender_id:
          type: string
        content:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        created_at:
          type: string
        read_at:
          type: string
          nullable: true
        deleted_at:
          type: string
          nullable: true
      required: [id, conversation_id, sender_id, created_at]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    AdminReport:
      type: object
      properties:
        id:
          type: string
        reporterId:
          type: string
        reporterUsername:
          type: string
          nullable: true
        targetUserId:
          type: string
          nullable: true
        targetUsername:
          type: string
          nullable: true
        category:
          type: string
        reportType:
          type: string
          nullable: true
        targetType:
          type: string
          nullable: true
        subject:
          type: string
          nullable: true
        message:
          type: string
        status:
          type: string
        actionTaken:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, reporterId, category, message, status, createdAt]
    AdminReportList:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/AdminReport"
        total:
          type: integer
      required: [reports, total]
    AdminVideoCall:
      type: object
      properties:
        id:
          type: string
        callerId:
          type: string
        calleeId:
          type: string
        status:
          type: string
        createdAt:
          type: string
        endedAt:
          type: string
          nullable: true
      required: [id, callerId, calleeId, status, createdAt]
    AdminReportDetail:
      type: object
      properties:
        report:
          $ref: "#/components/schemas/AdminReport"
        reporter:
          allOf:
            - $ref: "#/components/schemas/PublicProfile"
          nullable: true
        target:
          allOf:
            - $ref: "#/components/schemas/PublicProfile"
          nullable: true
        targetState:
          type: string
          enum: [active, suspended, banned]
        targetSuspendedUntil:
          type: string
          nullable: true
        targetReportCount:
          type: integer
        reportedMessage:
          allOf:
            - $ref: "#/components/schemas/Message"
          nullable: true
        messageContext:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        call:
          allOf:
            - $ref: "#/components/schemas/AdminVideoCall"
          nullable: true
        targetPhotos:
          type: array
          items:
            $ref: "#/components/schemas/ProfilePhoto"
      required: [report, targetState, targetReportCount, messageContext, targetPhotos]
    ResolveReportRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        note:
          type: string
          maxLength: 2000
        action:
          type: string
          enum: [none, suspend, ban, remove_photo, dismiss]
      required: [status]
    SuspendUserRequest:
      type: object
      properties:
        durationDays:
          type: integer
          description: Days until auto-lift; omit for indefinite
        reason:
          type: string
          maxLength: 500
    BanUserRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
    AdminFlag:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
          nullable: true
        reason:
          type: string
        detail:
          type: string
          nullable: true
        count:
          type: integer
        status:
          type: string
        createdAt:
          type: string
      required: [id, userId, reason, count, status, createdAt]
    AdminFlagList:
      type: object
      properties:
        flags:
          type: array
          items:
            $ref: "#/components/schemas/AdminFlag"
        total:
          type: integer
      required: [flags, total]
    ReviewFlagRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, reviewed, dismissed]
      required: [status]
    AdminSummary:
      type: object
      properties:
        openReports:
          type: integer
        openFlags:
          type: integer
        suspended:
          type: integer
        banned:
          type: integer
      required: [openReports, openFlags, suspended, banned]