# lib/api-zod/src/generated/types/adminFlag.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminFlag {
  id: string;
  userId: string;
  username?: string | null;
  reason: string;
  detail?: string | null;
  count: number;
  status: string;
  createdAt: string;
}