# lib/api-zod/src/generated/types/adminReportList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminReport } from './adminReport';

export interface AdminReportList {
  reports: AdminReport[];
  total: number;
}