# lib/api-zod/src/generated/types/banUserRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BanUserRequest {
  /** @maxLength 500 */
  reason?: string;
}